pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brahmx-sdg"
version = "0.1.0"
description = "Sovereign science-focused synthetic data generation and distillation platform"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [
    { name = "ZenteiQ Aitech Innovations Pvt. Ltd.", email = "legal@zenteiq.ai" }
]
keywords = ["llm", "synthetic-data", "distillation", "science", "verification", "sovereign"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Private :: Do Not Upload",
]

dependencies = [
    # LLM clients — OpenAI API (placeholder until TPU vLLM is ready)
    "openai>=1.30",
    "httpx>=0.27",
    # Retrieval
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.4",
    "numpy>=1.26",
    # Symbolic math
    "sympy>=1.12",
    # Units
    "pint>=0.23",
    # Schema validation
    "jsonschema>=4.22",
    "pydantic>=2.7",
    # Data formats
    "pyarrow>=15.0",
    "pandas>=2.2",
    # Utils
    "python-dotenv>=1.0",
    "typer>=0.12",
    "rich>=13.7",
    "pyyaml>=6.0",
    # Observability
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.10",
    "ruff>=0.4",
    "pre-commit>=3.7",
    "types-PyYAML",
]
kubeflow = [
    "kfp>=2.7",
    "google-cloud-aiplatform>=1.45",
]
training = [
    # MaxText deps installed separately in training containers
]

[project.scripts]
brahmx-sdg = "brahmx_sdg.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
include = ["brahmx_sdg*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "TCH", "ANN"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=brahmx_sdg --cov-report=term-missing"
